ZERO-81 [Rename]: develop 브랜치 파일구조 변경
